+++ Libplanet.RocksDBStore/RocksDBStore.cs
@@ -20,7 +20,7 @@ namespace Libplanet.RocksDBStore
     /// This stores data in the RocksDB.
     /// </summary>
     /// <seealso cref="IStore"/>
-    public class RocksDBStore : BaseStore
+    public class RocksDBStore : BaseBlockStatesStore
     {
         private const string BlockDbName = "block";
         private const string TxDbName = "tx";
@@ -649,16 +649,11 @@ namespace Libplanet.RocksDBStore
             }
         }
 
-        public override Tuple<HashDigest<SHA256>, long> LookupStateReference<T>(
+        public override Tuple<HashDigest<SHA256>, long> LookupStateReference(
             Guid chainId,
             string key,
-            Block<T> lookupUntil)
+            long lookupUntilBlockIndex)
         {
-            if (lookupUntil is null)
-            {
-                throw new ArgumentNullException(nameof(lookupUntil));
-            }
-
             if (_lastStateRefCaches.TryGetValue(
                     chainId,
                     out LruCache<string, Tuple<HashDigest<SHA256>, long>> stateRefCache)
@@ -668,14 +663,14 @@ namespace Libplanet.RocksDBStore
             {
                 long cachedIndex = cache.Item2;
 
-                if (cachedIndex <= lookupUntil.Index)
+                if (cachedIndex <= lookupUntilBlockIndex)
                 {
                     return cache;
                 }
             }
 
             Tuple<HashDigest<SHA256>, long> stateRef =
-                IterateStateReferences(chainId, key, lookupUntil.Index, null, limit: 1)
+                IterateStateReferences(chainId, key, lookupUntilBlockIndex, null, limit: 1)
                 .FirstOrDefault();
 
             if (stateRef is null)