+++ Libplanet/Net/Swarm.cs
@@ -546,6 +546,7 @@ namespace Libplanet.Net
                 : new BlockChain<T>(
                     BlockChain.Policy,
                     _store,
+                    _store as IStateStore,
                     Guid.NewGuid(),
                     BlockChain.Genesis,
                     false);
@@ -1448,7 +1449,8 @@ namespace Libplanet.Net
                         break;
                     }
 
-                    if (reply is RecentStates recentStates)
+                    if (reply is RecentStates recentStates
+                        && _store is IBlockStatesStore blockStatesStore)
                     {
                         if (recentStates.Missing)
                         {
@@ -1495,7 +1497,11 @@ namespace Libplanet.Net
                                     .ToImmutableHashSet();
                                 if (_store.GetBlockIndex(hash) is long index)
                                 {
-                                    _store.StoreStateReference(chainId, stateKeys, hash, index);
+                                    blockStatesStore.StoreStateReference(
+                                        chainId,
+                                        stateKeys,
+                                        hash,
+                                        index);
                                 }
                             }
 
@@ -1507,7 +1513,7 @@ namespace Libplanet.Net
                                 cancellationToken.ThrowIfCancellationRequested();
                                 IImmutableDictionary<string, IValue> states =
                                     pair.Value.ToImmutableDictionary();
-                                _store.SetBlockStates(pair.Key, states);
+                                blockStatesStore.SetBlockStates(pair.Key, states);
                             }
 
                             progress?.Report(new StateDownloadState()