+++ CHANGES.md
@@ -74,6 +74,19 @@ To be released.
     `Peer(PublicKey)` constructor.  [[#949]]
  -  Replaced `BoundPeer(PublicKey, DnsEndPoint, AppProtocolVersion)` constructor
     with `Peer(PublicKey, DnsEndPoint)` constructor.  [[#949]]
+ -  Extracted `IStore`'s some methods dedicated to block states into `IBlockStatesStore`. [[#950]]
+     -  `ListStateKeys()` method.
+     -  `ListAllStateReferences()` method.
+     -  `LookupStateReference()` method.
+     -  `IterateStateReferences()` method.
+     -  `StoreStateReference()` method.
+     -  `ForkStateReferences()` method.
+     -  `GetBlockStates()` method.
+     -  `SetBlockStates()` method.
+     -  `PruneBlockStates()` method.
+ -  The signature of `IStore.LookupStateReference<T>(Guid, string, Block<T>)` method was
+    changed to `LookupStateReference(Guid, string, long)`.  [[#950]]
+ -  Added `IStateStore`-typed `stateStore` to `BlockChain<T>` constructor.  [[#950]]
 
 ### Backward-incompatible network protocol changes
 
@@ -138,6 +151,10 @@ To be released.
  -  Added `Reorged` event on `BlockChain<T>`.  [[#945]]
  -  Added `ReorgedEventArgs` class.  [[#945]]
  -  Added `Swarm<T>.AppProtocolVersion` property.  [[#949]]
+ -  (Libplanet.RocksDB) `RocksDBStore` became to implement `IBlockStatesStore`.  [[#950]]
+ -  `DefaultStore` became to implement `IBlockStatesStore`.  [[#950]]
+ -  Added `IStateStore` interface.  [[#950]]
+ -  Added `IBlockStatesStore` interface.  [[#950]]
 
 ### Behavioral changes
 
@@ -232,6 +249,7 @@ To be released.
 [#945]: https://github.com/planetarium/libplanet/pull/945
 [#946]: https://github.com/planetarium/libplanet/pull/946
 [#949]: https://github.com/planetarium/libplanet/pull/949
+[#950]: https://github.com/planetarium/libplanet/pull/950
 [sleep mode]: https://en.wikipedia.org/wiki/Sleep_mode